Move Keras's get_regularization to a dedicated factory file
